Split the trait and implementation of S3ObjectExists
