+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/storage/services/ObjectExists.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.archive.common.storage.services
+
+import uk.ac.wellcome.storage.{ObjectLocation, StorageError}
+
+trait ObjectExists {
+  def exists(objectLocation: ObjectLocation): Either[StorageError, Boolean]
+}
