+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/storage/services/s3/S3ObjectExistsTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.storage.services
+package uk.ac.wellcome.platform.archive.common.storage.services.s3
 
 import org.scalatest.EitherValues
 import org.scalatest.funspec.AnyFunSpec
